more senseable locations for new components
